% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{forecast.lgt}
\alias{forecast.lgt}
\title{produce forecasts}
\usage{
\method{forecast}{lgt}(object, h = ifelse(frequency(object$x) > 1, 2 *
  frequency(object$x), 10), level = c(80, 95), NUM_OF_TRIALS = 2000,
  MIN_VAL = 0.001, MAX_VAL = 1e+38, ...)
}
\arguments{
\item{object}{lgt object}

\item{h}{Forecasting horizon (10 for annual and 2*periods otherwise)}

\item{level}{Confidence levels for prediction intervals a.k.a. coverage percentiles. Beween 0 and 100.}

\item{NUM_OF_TRIALS}{Number of simulations to run. Suggested rannge (1000,5000), but it may have to be higher for good coverage of very high levels, e.g. 99.8.}

\item{MIN_VAL}{Minimum value the forecast can take. Must be positive.}

\item{MAX_VAL}{Maximum value the forecast can take.}

\item{...}{description}
}
\value{
returns a forecast object compatible with the forecast package
}
\description{
This function produces forecasts from a model
}
\author{
bergmeir
}
