% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{lgt}
\alias{lgt}
\title{lgt class}
\usage{
lgt(y, lgtmodel, params, paramMean, seasonality)
}
\arguments{
\item{y}{the time series data}

\item{lgtmodel}{type of lgtmodel selected}

\item{params}{list of parameters}

\item{paramMean}{mean of each parameter}

\item{seasonality}{number of seasons, 1 for annual}
}
\value{
lgt instance
}
\description{
a constructor function for the "lgt" class
}
