% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlgt-package.R
\docType{package}
\name{Rlgt2-package}
\alias{Rlgt2-package}
\alias{Rlgt2}
\title{Getting started with the Rlgt2 package}
\description{
An implementation of LGT and SGT models as described in ...
}
\examples{
x <- 1
}
\author{
Slawek Smyl \email{slaweks@hotmail.co.uk} 

Christoph Bergmeir \email{christoph.bergmeir@gmail.com}
}
\keyword{exponential}
\keyword{forecasting,}
\keyword{smoothing}
