% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{lgt}
\alias{lgt}
\title{lgt class}
\usage{
lgt(y, lgtmodel, params, paramMean, seasonality)
}
\arguments{
\item{y}{desc 1}

\item{lgtmodel}{desc2}

\item{params}{desc3}

\item{paramMean}{desc4}

\item{seasonality}{desc5}
}
\value{
lgt instance
}
\description{
a constructor function for the "lgt" class
}
